Consolidate overlay e2e and reconnect harness into one runner
