+++ README_TESTING.md
@@ -4,7 +4,7 @@ Detailed testing notes are in `docs/README_TESTING.md`.
 
 This now includes dedicated sections for:
 - `tests/integration/test_overlay_e2e.py`
-- `tests/integration/test_overlay_e2e_reconnect.py`
+- `tests/integration/test_overlay_e2e.py --mode reconnect`
 
 For both suites, the doc describes:
 - how to start the suite,
